components/networked.js: this.data can be undefined
Sometimes the `this.data` value is undefined if there is a high
load of messages coming in, such as when a VR headset joins.
This causes the other avatar to disappear from the screen,
even though it still appears in the occupants array and
the other player can still see this one.

Signed-off-by: Trammell Hudson <hudson@trmm.net>